rename userPrompt.py to utils.py
